# perf: replace serde_json with simd_json